% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsc_at_x.R
\name{bsc_at_x}
\alias{bsc_at_x}
\title{Evaluates the functions \eqn{b} and \eqn{s} at \eqn{x}}
\usage{
bsc_at_x(x, bsc.list)
}
\arguments{
\item{x}{a value or a vector of x at which the functions \eqn{b} and \eqn{s}
are to be evaluated.}

\item{bsc.list}{A list produced by the function
\code{\link{bsc_ciuupi_sinc}}.}
}
\value{
Returns the dataframe containing \eqn{x} and the corresponding values of the
functions \eqn{b} and \eqn{s}.
}
\description{
Evaluates the functions \eqn{b} and \eqn{s} at \eqn{x} using a given list of
optimal knots \eqn{b} and \eqn{s_c}, as produced by the function
\code{\link{bsc_ciuupi_sinc}}.
}
\examples{
rho <- 0.6
alpha <- 0.05
opt.ciuupi <- bsc_ciuupi_sinc(rho, alpha)

x.seq <- seq(-10,10, length.out=1000)
bsc.x <- bsc_at_x(x.seq, opt.ciuupi)

}
\references{
\itemize{
  \item \href{https://doi.org/10.1016/j.jspi.2009.03.018}{Kabaila, P.
  & Giri, K. (2009). Confidence intervals in regression utilizing prior
  information. Journal of Statistical Planning and Inference 139, 3419-3429.}
  \item \href{https://cran.r-project.org/package=ciuupi}{ciuupi 1.2.3}.
}
}
\author{
A. Perera and P. Kabaila
}
