% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_standard.R
\name{ci_standard}
\alias{ci_standard}
\title{Computes the standard \eqn{1-\alpha} standard confidence interval}
\usage{
ci_standard(a, X, y, alpha, sig = NULL)
}
\arguments{
\item{a}{A specified nonzero \eqn{p}-vector which is used to define the parameter
of interest as \eqn{\theta=a^\top\beta}}

\item{X}{The \eqn{n \times p} design matrix}

\item{y}{The response vector of length \eqn{n}}

\item{alpha}{The desired minimum coverage probability
is 1\eqn{-}\eqn{\alpha}}

\item{sig}{Standard deviation of the random error. If not provided, then the
function will estimate the standard deviation under the condition that
\eqn{n-p \ge 30}.}
}
\value{
The lower and upper limit of the standard confidence interval for \eqn{\theta}
}
\description{
When \eqn{\sigma} is known this module computes the standard \eqn{1-\alpha}
confidence interval for the paramter of interest \eqn{\theta}.
If \eqn{\sigma} is not provided then the module will estimate the value of
\eqn{\sigma} using the provided data, given that \eqn{n-p \ge 30}.
}
\details{
Consider the linear regression model:
\deqn{Y=X\beta + \epsilon}
where \eqn{Y} is a random \eqn{n}-vector, \eqn{X} is the \eqn{n \times p}
design matrix, \eqn{\beta} is the unknown parameter \eqn{p}-vector and
\eqn{\epsilon} is the random error with a \eqn{N(0, \sigma^2 I)} distribution.
Suppose that the parameter of interest is:
\deqn{\theta = \bm{a}^{\top} \bm{\beta},}
where \eqn{\bm{a}} is a specified
nonzero \eqn{p}-vector.

For a given response vector \eqn{Y} and desing matrix \eqn{X} this function
computes the standard confidence interval for the parameter of interest \eqn{\theta}
that has a minimum coverage probability \eqn{1-\alpha}.
}
\examples{
y <- c(87.2, 88.4, 86.7, 89.2)
x1 <- c(-1, 1, -1, 1)
x2 <- c(-1, -1, 1, 1)
X <- cbind(rep(1, 4), x1, x2, x1*x2)
a <- c(0, 2, 0, -2)
ci_standard(a, X, y, 0.05, sig = 0.8)

}
\references{
\itemize{
  \item \href{https://doi.org/10.1016/j.jspi.2009.03.018}{Kabaila, P.
  & Giri, K. (2009). Confidence intervals in regression utilizing prior
  information. Journal of Statistical Planning and Inference 139, 3419-3429.}
  \item \href{https://cran.r-project.org/package=ciuupi}{ciuupi 1.2.3}.
}
}
\author{
P.Kabaila
}
