% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciuupi_observed_value.R
\name{ciuupi_observed_value}
\alias{ciuupi_observed_value}
\title{Computes the confidence interval that utilizes uncertain prior information
for a given observed vector \eqn{Y}}
\usage{
ciuupi_observed_value(a, c, X, alpha, bsc.list, t, y, sig = NULL)
}
\arguments{
\item{a}{The \eqn{p} vector \eqn{\bm{a}} used to specifies the parameter of
interest (\eqn{\theta=\bm{a}^\top \beta})}

\item{c}{The \eqn{p} vector \eqn{\bm{a}} used to specifies the uncertain prior
information (\eqn{\tau=\bm{c}^\top \beta}). The uncertain prior information is
that \eqn{\tau=0}.}

\item{X}{The \eqn{n \times p} design matrix.}

\item{alpha}{The desired minimum coverage probability
is 1\eqn{-}\eqn{\alpha}}

\item{bsc.list}{A list produced by the function
\code{\link{bsc_ciuupi_sinc}}.}

\item{t}{The numerical value used to specify the uncertain prior information
\eqn{\tau=\bm{c}^\top \beta}. The uncertain prior information is
that \eqn{\tau=0}.}

\item{y}{The response vector of length \eqn{n}}

\item{sig}{Standard deviation of the random error. If not provided, then the
function will estimate the standard deviation under the condition that
\eqn{n-p \ge 30}.}
}
\value{
The lower and upper limits of the confidence interval for \eqn{\theta} that
utilizes the uncertain prior information.
}
\description{
For known \eqn{\sigma} this function computes the confidence interval that
utilizes uncertain prior information for the
parameter of interest for a given observed vector \eqn{Y}.
If \eqn{\sigma} is not provided this function estimates the value of
\eqn{\sigma} provided that \eqn{(n-p)>30}.
}
\details{
Consider the linear regression model:
\deqn{Y=X\beta + \epsilon}
where \eqn{Y} is a random \eqn{n}-vector, \eqn{X} is the \eqn{n \times p}
design matrix, \eqn{\beta} is the unknown parameter \eqn{p}-vector and
\eqn{\epsilon} is the random error with a \eqn{N(0, \sigma^2 I)} distribution.
Suppose that the parameter of interest is:
\deqn{\theta = \bm{a}^{\top} \bm{\beta},}
where \eqn{\bm{a}} is a specified
nonzero \eqn{p}-vector.
Let
\deqn{\tau = \bm{c}^{\top} \bm{\beta} - t,}
where \eqn{\bm{c}} is a
specified nonzero \eqn{p}-vector that is linearly independent of \eqn{\bm{a}}
and \eqn{t} is a specified number.
Suppose that previous experience with similar data sets and/or expert
opinion and subject-matter background suggests that \eqn{\tau = 0}.

For a given response vector \eqn{Y} and desing matrix \eqn{X} this function
computes the confidence interval for the parameter of interest \eqn{\theta}
that has a minimum coverage probability \eqn{1-\alpha},
that utilizes the uncertain prior information \eqn{\tau=0}
}
\examples{
a <- c(0, 2, 0, -2)
c <- c(0, 0, 0, 1)
x1 <- c(-1, 1, -1, 1)
x2 <- c(-1, -1, 1, 1)
X <- cbind(rep(1, 4), x1, x2, x1*x2)
alpha <- 0.05
t <- 0
y <- c(87.2, 88.4, 86.7, 89.2)
sig <- 0.8
rho <- 0.6
alpha <- 0.05
opt.ciuupi <- bsc_ciuupi_sinc(rho, alpha)
ciuupi_observed_value(a, c, X, alpha, opt.ciuupi, t, y, sig=sig)

}
\references{
\itemize{
  \item \href{https://doi.org/10.1016/j.jspi.2009.03.018}{Kabaila, P.
  & Giri, K. (2009). Confidence intervals in regression utilizing prior
  information. Journal of Statistical Planning and Inference 139, 3419-3429.}
  \item \href{https://cran.r-project.org/package=ciuupi}{ciuupi 1.2.3}.
}
}
\author{
P.Kabaila and A.Perera
}
