% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_at_gamma.R
\name{cp_at_gamma}
\alias{cp_at_gamma}
\title{Compute the coverage probability of the confidence interval that utilizes the
uncertain prior information at a given vector of \eqn{\gamma}.}
\usage{
cp_at_gamma(gamma.vec, bsc.list)
}
\arguments{
\item{gamma.vec}{The value of \eqn{\gamma} or a vector of \eqn{\gamma} that
at which the coverage probability is needed to be evaluated.}

\item{bsc.list}{A list produced by the function
\code{\link{bsc_ciuupi_sinc}}.}
}
\value{
Computes the coverage probability of the CIUUPI for a given value or
vector of \eqn{\gamma}.
}
\description{
Evaluate the coverage probability of the confidence interval that utilizes
uncertain prior information (CIUUPI) at a given \eqn{\gamma}. This function
uses the output produced by \code{\link{bsc_ciuupi_sinc}}.
}
\examples{
rho <- 0.6
alpha <- 0.05
bsc.list <- bsc_ciuupi_sinc(rho, alpha)

gamma.vec <- seq(0,10, length.out=10)
cp.gamma <- cp_at_gamma(gamma.vec, bsc.list)

}
\references{
\itemize{
  \item \href{https://doi.org/10.1016/j.jspi.2009.03.018}{Kabaila, P.
  & Giri, K. (2009). Confidence intervals in regression utilizing prior
  information. Journal of Statistical Planning and Inference 139, 3419-3429.}
  \item \href{https://cran.r-project.org/package=ciuupi}{ciuupi 1.2.3}.
}
}
\author{
A.Perera
}
