% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_b.R
\name{plot_b}
\alias{plot_b}
\title{Plot the graph of \eqn{b} function}
\usage{
plot_b(bsc.list)
}
\arguments{
\item{bsc.list}{A list produced by the function
\code{\link{bsc_ciuupi_sinc}}.}
}
\value{
The plot of the odd function \eqn{b} used in the specification of the
confidence interval.
}
\description{
This module plots the odd function \eqn{b} that is used to specify the
confidence interval for \eqn{\theta} that utilizes the uncertain prior
information.
}
\examples{
rho <- 0.6
alpha <- 0.05
bsc.list <- bsc_ciuupi_sinc(rho, alpha)
plot_b(bsc.list)


}
\references{
\itemize{
  \item \href{https://doi.org/10.1016/j.jspi.2009.03.018}{Kabaila, P.
  & Giri, K. (2009). Confidence intervals in regression utilizing prior
  information. Journal of Statistical Planning and Inference 139, 3419-3429.}
  \item \href{https://cran.r-project.org/package=ciuupi}{ciuupi 1.2.3}.
}
}
\author{
A. Perera and P. Kabaila
}
